+++ packages/emojibase/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.2.0
+* Updated `fetchFromCDN` to cache responses in session storage.
+* Removed tests from distribution files.
+
 # 1.1.0
 * Added new `stripHexcode` function to the `emojibase` package.
 * Fixed Flowtype definitions.